Re-Organize components into section folders
