+++ seeit/src/components/headerBar/DeleteAll/style.css
