+++ seeit/src/components/results/ResultCards/style.css
