+++ seeit/src/components/results/ResultRow/style.css
