+++ seeit/src/components/search/SearchSection/style.css
