+++ seeit/src/components/ui/RatingStars/style.css
