+++ seeit/src/components/ui/ToolIcon/ToolIcon.js
@@ -2,17 +2,17 @@ import React from 'react';
 import './style.css';
 
 
-const ToolIcon = ({ clickHandler, iconClass, circle, path }) => {
+const ToolIcon = (props) => {
 
     return (
 
         <svg
-            onClick={clickHandler}
-            className={iconClass}
+            onClick={props.clickHandler}
+            className={props.iconClass}
             viewBox="0 0 60 60"
             xmlns="http://www.w3.org/2000/svg">
-            {circle}
-            {path}
+            {props.circle}
+            {props.path}
         </svg>
     )
 }
